feat: 선후배에 따라 분기처리
