+++ src/pages/promiseList/commons/PromiseTap.tsx
@@ -1,4 +1,5 @@
 import styled from '@emotion/styled';
+import ProfileContainer from './ProfileContainer';
 
 const PromiseTap = () => {
   return (
@@ -8,6 +9,13 @@ const PromiseTap = () => {
         <TapText $isActive={false}>예정 약속</TapText>
         <TapText $isActive={false}>지난 약속</TapText>
       </TapContainer>
+      <ProfilWrapper>
+        <ProfileContainer name="도리야끼다요" userRole="SENIOR" type="waitingAppointments" />
+        <ProfileContainer name="도리야끼다요" userRole="SENIOR" type="waitingAppointments" />
+        <ProfileContainer name="도리야끼다요" userRole="SENIOR" type="waitingAppointments" />
+        <ProfileContainer name="도리야끼다요" userRole="SENIOR" type="waitingAppointments" />
+        <ProfileContainer name="도리야끼다요" userRole="SENIOR" type="waitingAppointments" />
+      </ProfilWrapper>
     </Wrapper>
   );
 };
@@ -15,6 +23,10 @@ const PromiseTap = () => {
 export default PromiseTap;
 
 const Wrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+
   width: 100vw;
 
   background-color: ${({ theme }) => theme.colors.grayScaleWhite};
@@ -29,7 +41,13 @@ const TapContainer = styled.div`
   padding: 1.5rem 4.7rem;
   border-bottom: 1px solid ${({ theme }) => theme.colors.grayScaleLG1};
 `;
+
 const TapText = styled.span<{ $isActive: boolean }>`
   ${({ theme }) => theme.fonts.Title1_SB_16};
   color: ${({ theme, $isActive }) => ($isActive ? theme.colors.grayScaleBG : theme.colors.grayScaleMG2)};
+`;
+
+const ProfilWrapper = styled.div`
+  width: 100vw;
+  padding: 2rem;
 `;
